Support disassembling generic functions and methods
In extlib.stub, method-code is added to extract compiled Scheme code
from a method.  It returns #f if the argument is a C-defined method.

The disasm procedure handles the case when the argument is a method
or a generic function.  If it is a generic function, disasm prints
disassemble of all methods.

The output format is provisional; it can be improved.