Finishing pass4 optimization (phase 1)
We lift independent and self-recursive closures to toplevel.
compile-toplevel-labmda is dropped, for a single toplevel closure
can become several toplevel defines because of lifted lambdas.
We should re-evaluate the effect of dropping compile-toplevel-lambda,
since it was originally introduced to reduce redundancy of the
compiled code.  Performance evaluation is also on the todo list.