Fixed a compiler bug that affects inline expansion
pass1/define-inline-finish has a code that relies on the fact
that every lambda form returns a fresh closure.  It was no longer
true since lambda lifting optimization was incorporated---if the
lambda form does not close local variables, then a single instance
of closure is created in the toplevel environment.

This caused mysterious failure in ext/util/queue.scm, in which
inline information is overwritten by subsequent define-inline form.

The effect is only visible with precompilation, since the affected
code uses the closure only for a placeholder to inline the procedure
during compilation of the single compile unit.  In the normal mode
of operation, every toplevel form consists of a compile unit, and
that's why this wasn't caught in the normal test cases.

The code since commit d6f23ad1 until this fix should not be used
to compile Gauche itself, since it yields a corrupt compiler.