Dropped the support of unofficial, obsoleted keyword argument spec
The obsoleted keyword argument spec, (var keyword init) was used
in some Gauche internal code.  This was never been documented,
and replaced in favor of CL-compatible ((keyword var) init)
in 0.9 release.  The support was kept for a while in order to
allow Gauche to be built with pre-0.9 release.  Now we're already
0.9.1, so the old form can be dropped safely.