Fixed change-class on class metaobjects
Changing metaclasses, that is, calling change-class on class
metaobjects hasn't been working, since the slots of the
class essential for the operation of the object system
wasn't carried over.  It is fixed.

And also a new mechanism to protect accidental modification
of those essential slots are in place (SCM_CLASS_MALLEABLE flag).
Basically, those slots can only be set during initialization
of the class.  At this moment change-class won't run initialize
method, so these slots can't be modified by change-class only.
(This might be changed later if we introduce reinitialize
method.)

A new method change-object-class-carry-over-slot? is added
to customize slots to be carried over; I still feel this is
a bit of kludge, so let's keep it unofficial for the time being.