Dropped the dependency on an implicit local variable in init function.
In the precomp-generated C source, a local variable ScmModule *mod
used to be defined to hold the module of the compiled source.  It's
ugly to depend on such implicit names, and also it was awkward to
handle the situation when multiple modules are defined in the
precompiled source.  Now, if you want to refer to the module inside
init functin, you can call Scm_CurrentModule().