user_eval_inner does no longer save/restore the signal mask
We don't see any reason to restore the mask.  From the commit logs,
this mask manipulation was originally kept in 2004-03-15 change;
it says, without restoring the mask, a hang occurred on cygwin.
We can no longer reproduce the hang, so we decided to drop this code.
It has now become a major slowdown factor.