Fixed the hazard around resetting owner slot of lazy pair.
When we throw an error if we detect the recursive forcing of
a lazy-pair, there should be an unwind handler that resets the
lp->owner in upstream, so we don't need to reset it before
calling Scm_Error().  The previous code had a hazard that
another thread grabs the lazy pair between we raise Scm_Error
and the error is catched by the unwind handler, and in that
case the owner slot would be overwritten.