Improved regexp matcher speed by skipping the input efficiently.
We calculate a set of characters that can possibly match at the
beginning of the regex (lookahead set, or laset).  When we
advance the input pointer, we check the character and if it
isn't in the laset we can skip matching from that position.
We also adopt a strategy that when /A+B/ fails to match, where
A is a char or charset, and A and B has distinct laset,
then we can skip the input as long as it matches A.  These
strategy improves the performance greatly in commonly occurring
patterns, especially when the input is long.

Some benchmark results:
http://blog.practical-scheme.net/shiro/20111021-regexp