Added support of zero-length mt-queue.
Zero-length queue cannot hold any values, but can be used to synchronize
producers and consumers---a thread can directly hand a value to another
thread.  THIS IS AN INCOMPATIBLE CHANGE: It used to be that
zero max-length meant to be unlimited length.  Now you have to
give #f to max-length to make the mtqueue unlimited length.