Fixed a bug that the internal value of lazy sequence could leak out.
The generator of the lazy sequence may yield two values, the second
of which being a new generator.  When this happens the vm's numVals
counter has set to 2.  If ForceLazyPair was a tail call from the user's
code, this numVals was propagated and the new generator value was
leaked out.