Fix the error when #?= is used in pre-compiled code
If code using #?= is precompiled, an error is raised when the code
is executed.  It's because #?= (debug-print) is expanded to calls
to debug-print-pre and debug-print-post at the compile time.  They
are unexported symbols in gauche.vm.debug, and not set to be
autoloaded.  When the precompiled code runs and gauche.vm.debug isn't
loaded explicitly, the runtime can't find these symbols and barfs.

For a quick (but not ideal) solution, we just export
debug-print-(pre|post) and set autoload.