Fixed regexp ^ and $ parsing within assertion group
To resolve ambiguity of '^' and '$', we should treat (?=...), (?!...),
(?<=...), (?<!...) and (?>...) as if they are stand-alone, no matter
where they appear in the containing regexp.  That is, '^' appearing at
the beginning of these constructs are BOL assertions and '$' at the
end are EOL.  This is consistent to Perl and Oniguruma.  NB: The code
that counted on the previous (buggy) behavior may break by this
change.