Fixed race condition in control.job
If we allow a job status to be modified by threads other than
the job runner, via job-mark-killed!, we need a mutex.  So we
added a 'cancellable' job flavor, which may not be waitable
but use mutex to change the status.  See the comment in control/job.scm
for the details.