Graceful thread termination
Terminating thread by pthread_cancel is too dangerous.  We added
two "more graceful" termiantion attempt before trying the brutal method.
First, we extend vm->stopRequest mechanism.  If the target thread is
exeuting in vm loop, it works.  Then, we use pthread_kill to signal
the target---if the target thread is in C routine or blocked in syscall,
it should work.  We reserver one signal for this purpose.  Finally,
if the target doesn't respond on the above two, we use pthread_cancel()
as the last resort.