Fixed VM stack corruption bug in Scm_ForceLazyPair
VM stack is corrupted when lazy-pair forcing occurs during a comined
instruction such as CAR-PUSH.  The original code pushed an extra
continuation frame during forcing to protect the stack.  If that
happened, CAR-PUSH pushes the result on top of the extra continuation.
However, the next instruction to be executed after this CAR-PUSH will
be RET, set by Scm__VMUnprotectStack, which discards the just pushed
value by CAR-PUSH.  The fix correctly adjusts the VM stack after
forcing is done.