Do not allow strings with NUL chars in it for pathnames etc
It may be a security hazard when a Scheme string with NUL chars
in it is passed to a system procedures such as open-output-file.
A new stub type <const-cstring-safe> should be used for such procedures
to do extra check.

We considered prohibiting strings with NUL in Scm_GetStringConst
altogether, but in rare cases C API may need to accept such strings,
so we decided having a separate stub type.