Improving rfc.json and parser.peg API
It became clear that we need to get "the rest of input" after
parsing somehow---port is unreliable after calling one parser on it,
since lseq can prefetch arbitrary number of characters.  One immediate
drawback is to parse multiple JSON exprs from input; the current
interface fails when the input has no whitespace between JSONs, since
the first character of the next JSON expr is prefetched by the
peg parser and discarded.

We modified peg-run-parser to return the lseq as the second
value, which can be used for subsequent parsing.

For rfc.json, we're not ready to expose internal parser mechanism
yet; we provide a higher-level API parse-json* to plug the defect.
We'll think more general API later.