Changed power-of-two? to twos-exponent to return more meaningful result
It is more useful for power-of-two? to return the exponent when
the input is 2^k, instead of just #t.  But then, it is not a simple
predicate so having the name 'power-of-two?' is against convention.
We used twos-exponent for the improved procedure, and renamed the
previous twos-exponent to twos-exponent-factor.  (I'm not totally
happy for these names, since the two seem confusing.)