Renamed random data generation APIs
They are generators - so we use plural names.  E.g. fixnums are
a generator to produce bunch of fixnums.  We also suffix the
APIs that creates primitive generators by '$' to avoid confusion
(e.g. 'fixnums' is a generator, 'integers$' is a procedure that
returns an integer generator).

For combinators we don't add suffix yet (e.g. 'lists-of').
For the consistency we should, but
(list-of$ (vector-of$ 3 (integers$ 10))) looks too
many $'s.  Let's see how it goes.