Support 'export' form without *.sci file
This is an internal matter---for precompiled extension modules, export
form will be included in *.sci file as a part of define-module form,
so the C part doesn't need to worry about exporting.  However, to build
the Gauche core, where we don't generate *.sci files, we do need exporting
in the C init function.