Recognize ellipses inserted by macro expansion
Macro generating macro may insert ellipses to output, which becomes
identifiers in current implementation.  We need to recognize them
as well.

This fix is not complete; we need proper free-identifier=? semantics
to compare inserted ellipses and the ellipsis the current matcher
is using.  We leave the fix to the new low-level hygienic macro
facility.