Start using C99.
As we discussed, it's probably time to allow C99 features, especially mixing
declarations and code, in Gauche souce code. Using C99 allows us to write more
readable code.

I think that MSVC is the only major C compiler that does not support C99,
but because (1) we are using MinGW for Windows support, and (2) MSVC 2013
started supporting most C99 features, I'd think that wouldn't block us from
migrating to C99.

Compiler may need -std=c99 or -std=gnu99 to turn on C99 support.
AC_PROG_CC_STDC will do that for you if you have autoconf 2.69 or later. We
already have the macro call in configure.ac.

This patch is to convert list.c to C99 as a starter. If it looks good, I'll do
the same thing on other files too.