Renamed $fmap to $lift
The name 'fmap' is taken from Haskell's Applicative, but it doesn't
fit, for 'map' strongly suggests something in line of
(a -> b), [a] -> [b].  Since this function lifts ordinary function
to monadic parser, that is, (a,... -> b), m a, ... -> m b,
'lift' sounds more appropriate.

Also we dropped $<<, for it is equivalent to $lift except it's a macro.
We think making $lift inlinable addresses the performance concern.
(We could do better for specializing to typical 2 or 3-argument
cases; later).

And also made $->rope etc. to take multiple parsers; we find it's
common to want to concatenate results of several parsers into
a rope.  This is experimental; we may revert if we see better
abstraction.