Prevent sign extension of char in string pointer accessors
When string pointers on a single-byte string retrieve a char
and it has code between 0x80 and 0xff, it was treated as a signed
char and its sign is extended, resulting negative ScmChar value.

They aren't noticeable with native encodings other than 'none', since
on other encodings having that byte range in strings means it's
a multibyte string and the code path was never executed.