In 'none' native encoding, let ucs->char return subst char for invalid range
It is pretty inconvenient when literal character notation beyond
0xff throws an error when the native encoding is 'none'; even the
program uses cond-expand, the error occurs before macro expansion
(at read stage).  So we let the reader return a substititon char #\?
in such case.

This may allow subtle bugs to slip into the code if a programmer
isn't careful enough to use cond-expand.  We think it's acceptable
since encoding='none' is deemed to be gone by 1.0 release.