Copy filehandle in sys-get-std-handle (os.windows)
When #<win:handle> Scheme object is GC-ed, associated file handle
is closed.  Sys-get-std-handle should copy the handle returned
by GetStdHandle, so that the original std handle remain intact
even the returned Scheme object is GC-ed.

Another possible solution is to add 'owner' flag in the Scheme-side
of the object.  For now, copying is easier; we'll see.