Make all built-in modules "provided"
This used to be done in load.c, but we had to maintain list of modules
manually.  Now we extract all the modules at the last stage of initialization
to mark provided.

The cond feature list (in src/core.c) and the predefined module list
(src/module.c) is also updated.