Enable support of symbol-keyword unification
It is turned off by default for now, to keep the backward
compatibility.  Setting environment variable GAUCHE_KEYWORD_IS_SYMBOL
turns on the feature.

In the legacy "keywords and symbols are disjoint" mode, keyword
are kept in a separate table, <keyword> class doesn't inherit
<symbol>, and keyword->name doesn't contain the leading colon.

In the new "keyword is subtype of symbol" mode, keywords are
kept in the symbol hashtable, <keyword> class inherits <symbol>,
and keyword->name contains the leading colon, for it's the
name of the symbol.

We'll flip the default after a while, and eventually eliminate
all the legacy code once we think exiting code are adapted.