Adjust make-comparator behavior on default procedures
Let make-comparator accept #t or #f for fallback behaviors, as
specified in srfi-114.  We modified ScmComparator struture to
support fallback compareFn/hashFn behavior (that is, if #f is
given to compare-proc at construction time,
comparator-comparison-procedure? should return #f, but
comparator-comparison-procedure should return a fallback proc.)
This technically breaks ABI compatibility but nobody should be
using comparator in 0.9.4 so it should be ok.