Fix macro transformer not to include reference to current-module
For AOT compilation we expand primitive and er macro transformer
into a form that refers to an expression that makes toplevel env
at runtime (see the comment in primitive-macro-transformer).  The
expanded form shouldn't have bare reference to (current-module),
though, for it may not be visible from the macro calling site.