util.dominator fix - we don't need to include (start start) in output
The algorithm uses doms[start] = start for the start condition, but
we don't need to include that relation in the output; a node is a
dominator of itself so we don't need to say the obvious.  Also all
other elements are (node immediate-dominator-node), so we don't
want the start node to be treated specially.