Introduced ScmWriteParameter
We keep write customization parameters, such as *print-level* in CL,
in a structure in VM for ease of access (accessing parameter slots for
every time their value is needed is tedious.)

The means to change those parameters yet to come.

Not all parameters are used yet.

The name "WriteParameter" is provisional; it can be confused with
the srfi-39 parameters.  We may change it later.