Recover the feature to print PVREF
PVREF is an internal object only used in the compiled macro template
in the current macro expander.  As we plan to replace the expander,
and since it never appears elsewhere, somehow we've dropped its printing
routine from write.c inadvertently; but we do need it when
we #define DEBUG_SYNRULE in macro.c.