text.line-edit - Enable multiline edit
If (~ ctx'input-continues) returns true, enters multiline edit mode.
C-p and C-n work as vertical cursor movement when it can move so,
and prev/next-history if the cursor is on the edge.  M-p and M-n are
bound to prev/next-history.
The current implementation doesn't work if the # of lines exceeds
the screen height.
The secondary prompt needs to be thought over.