gauche.generator - Changed gtake optional arg spec for srfi-121
This is incompatible change.  Make gtake optional arg work the
same as srfi-121's gtake.  There could be some cases that this
break existing code; see the comment.  The easiest way is to
use gtake*, or to fully specify two optional args in order to
make it work with both older and newer Gauche.