gauche.cgen.unit - Change cgen-emit-h to always generate file
This reverts the previous change, and makes cgen-emit-h always
create .h file.  Also makes cgen-unit-h-file always return a
string.  This behavior aligns the document.

Rationale: There are cases that .h file is not needed, but it is
confusing that cgen-emit-h may or may not generate it.  If the
program calls cgen-emit-h, it is asking to create .h file explicitly,
so we should create it even if we know it's empty.