Add 'sealed' flag in ScmModule
Sealed modules won't allow modifying their bindings nor importing
further modules.  It may be generally useful (sort of package-lock
in Allegro CL), but the immediate reason is to address the problem
of 'require'; its fix will come in a separate commit.
Since we don't know general use cases for this feature, we make
it somewhat hidden for the time being (i.e. %seal-module! is in
gauche.internal).