Fix build on Mac OSX and make ioctls actually works.
The build fails since there is no ifr_netmask in struct ifreq on OSX.
In addition, some ioctl like SIOCGIFNETMASK doesn't work because
the actual value of SIOCGIFNETMASK on OSX is 0xc0206925 and
passing it as int and using it on ioctl, which expects unsigned long,
causes unwanted integer promotion.

There is still build warnings though.