Fix compiler to avoid global bindings for lifted lambdas in pass4
Pass4 used to create new global bindings for lifted lambdas, but
it was ambiguous that in which module the new bindings to be created;
the forms like define-module, with-module and select-module are all
dissolved in pass1 and we don't have a proper info at pass4.

We now create a new letrec surrounding the original expr, and use
local bindings for lifted lambdas.