Allow abbreviation in REPL toplevel commands
Based on https://github.com/shirok/Gauche/pull/200

I'm not entirely sure if this is good or not, since: (1) This
may give unexpected behavior when an unfamiliar is exploring.
For example, one may thing there might be a "print" command and
try ",p" but what she sees is the directory name printed.  It
needs some extra steps to find out which command she really
executed.  The pwd command is innocent enough, but what if it
is a command with some side effects?
(2) Sometimes explicit abbreviation is better.  If we add "doc"
command, we still want to reserve ",d" for "describe".

Anyway we'll give a try and see how it works out.