gauche.cgen.literal - Support for extended pairs
We don't serialize pair attributes automatically, for there may
be unexpected info attached that may not be serializable or
unnecessary.  However, sometimes we want to generate extended
pair in the precompiled data.  To do so, we provide a special
object that works as a proxy of extended pair, and can be
serialized into precompiled file.