Clear DESTDIR when creating symlinks to so in place
The makeverslink script refers to DESTDIR so that it can create
symlinks to the installed *.so, in case files are installed
under $DESTDIR for packaging.  The script also called during
build to create symlinks inside src/, at that point we assume
DESTDIR isn't set.  The assumption breaks if it is set externally.

This addresses https://github.com/shirok/Gauche/issues/202