Simplify source info handling in precompiled closures
Now that (~ <compiled-code>'sourceinfo) slot has signature with
source location info attached, we can just serialize it.  It will
be carried over to <closure> by Scm_MakeClosure.

The drawback is that the source location info won't be avaialble
for procedures in src/*.scm, since they are processed by 0.9.4.
But that's just a transient thing.