text.line-edit - Emit CRLF at committing input
We usually commit with ENTER key, and even if not, it's sensible
that the response of program starts at the fresh line.  This behavior
is also consistent with the ordinary read-line (so that we can just
replace it for read-line/edit).  Note that, as read-line does,
the final newline isn't added to the return value.