Another fix on static build wrt. avoid embedding identifiers
Reader macro may inserts identifiers, and it can't be written out
portably.  Currently we only have string interpolation that does
so, and we know Gauche bundled code doesn't rely on hygienity of
them, thus we simply strip identifier info.  In future, though,
we should be careful when more general reader macros are introduced.