Changed string interpolation expansion
This abandons hygienity we introduced several weeks ago.  Inserting
identifiers by reader causes unexpected consequences, and we decided
it could do more harm than good.  String interpolation syntax is
Gauche-specific anyway, so it won't be unreasonable to ask it to
be used when gauche namespace is visible.

Now the reader macro expands interpolation to a call of macro
'string-interpolate*'.  It is the only identifier that needs to be
lexically visible.  The remaining expansion is done at macro
expansion phase, which takes care of hygiene, and we'll have flexibility
to change how the final expansion can be.