Add SCM_PORT_OUTPUT_TRANSIENT mode in port
The port API accepts it in DIR argument; internally it's separated
to SCM_PORT_OUTPUT direction and SCM_PORT_TRANSIENT flag.
Transient output is specifically for dumping data temporarily
during the process execution; the data isn't supposed to survive
the process, so we don't register it to the table to flush upon
process exit.

Also cleaned up storing fd in file buffered port.