Refactoring procedural inliner - step 1
It makes more sense for procedural inliners (defined by
define-builtin-inliner) to work on IForms instead of the original
S-expr; by that we can separate S-expr to IForm conversion from
inline expansion, and also it allows us to call inliner in later
passes.

As a side effect, the change made builtin inliners much simpler.

Currently the compiler macro is implemented on top of procedural
inliner, but it needs to work on S-expr instead of IForms; so this
commit doesn't touch the calling part of inliner yet.  We'll split
compiler macro invocation next.