Fix positional matching of records in util.match
The '$' pattern in util.match uses position to match slots in objects.
When the object is a record, it is naturally matches in the order
of declared fields.  However, when a record used inheritance, class-slots
orders the slots in derived record before inherited slots.  That order
is pretty unintuitive as it doesn't match the argument order of constructor,
and useless as it prevents from matching instance of subclass as its
superclass.

Unfortunately, the document stated the positional matching is done in
the order of class-slots.  So this change breaks publicized API.  Still
we believe the fixed order is right.