Fix record positional match
The <slot-accessor> object only handles instances of the associated class,
but not their subclasses---because the instance layout of subclass can be
anything.

We ended up using instance-slot-ref.  It's not optimal to directly access
instance, but we *know* the specifics of records (notably, it won't have
any fancy slots) so it should work.  We wrap it with closure at the
class initialization, so we won't accdentally access incorrect slots.