Change return value of Scm_{HashTable|TreeMap}Set
We see the new spec is more consistent and it provides more
info to the caller.  Unfortunately this is incompatible change;
we bet these functions are almost always used just for side-effects,
but if there's code that uses their return value, it'll be
broken; let us know if there's such a case.