Experimentally add -v VERSION option to gosh
We keep older versions of gauche installation by default, so theoretically
the user can use different versions of gosh in parallel.  It has been
cumbersome since the version-dependent path to the executable is rather
long.

With this version, the 'master' (usually the latest) gosh can exec's
the specified version of gosh if it is installed.

There's an overhead for looking the binary and exec() call, so it is
not recommended to use this feature for frequently used script.
It will be handy, though, for testing scripts with different versions
of Gauche.