sxml: do not end html5 empty tags with "/>"
HTML5 does not seem to accept all tags ending with "/>" like XML (or
HTML4) [1]. Those of "void elements" will accept and ignore the
"/". Foreign tags (MathML and SVG) do recognize that. For other tags, it
seems ignored as well and result tags remain open.

The code is updated so that we still leave void elements open (which is
fine; closing tags are optional for them). Otherwise, we immediately
close the tag.

[1] https://www.w3.org/TR/html5/syntax.html#start-tags