Trigger load-hook with paths beginning with '@'
Previously, we trigger load hooks when a load path prefix points to
a regular file or "".  It turnes out we need to deal with relative
paths from the file loaded via load hooks, so we need to keep source
path info on files loaded not from files.

This mechanism covers the original use of "" (by adding "@" in the
load path), so we deprecate the previous use of "".