sxml: fix ssax parser to trim whitespace according to xml:space
XML standard [1] specifies that if a tag has xml:space="preserve" and
it only contains whitespaces then all the whitespaces should be
preserved.

The current SSAX parser does parse and pass xml:space around (child
nodes inherit this attribute from the parent node) but it does not
actually follow this rule.

Fix this by checking for the presence of xml:space then decide whether
to drop whitespaces. The fix is not perfect, we should use
preserve-ws? that's propagated throughout the nodes. Unfortunately we
don't have access to it at this point of dropping whitespace.

[1] https://www.w3.org/TR/xml/#sec-white-space