Fix process connection.  Remove 'detached' flag.
It is more consistent to make process cleanup in shutdown rather than close.
The subprocess reading from stdin usually exits when the input is closed,
and that happens when we shutdown the process connection.  There's no
reason that we want to keep process alive after that, so we can take
more steps to ensure the process is really terminated.

OTOH, connection-close merely closes the ports; if forked process is
still connecting to the process, the process keep working.