Removed pre-package-ext make target
It was originally created to pre-generate C files in ext/ directories,
because with Windows/VC++ it was complicated to use freshly built gosh
to bulid extensions.  So we used to run 'make pre-generate-ext' on Linux
and carry over generated c files on Windows to compile.

However, we've changed the extension build system so that the we would
no longer keep intermediate C files, thus pre-package-ext target became
obsolete.