parser.peg - Tweak error message from $fail
If $fail is the cause of the parse error, we treat its message
as the definitive source of error message.  With this, the patten like this:
  ($or A-parser B-parser C-parser ($fail "Expected A, B or C"))
Would yield a error mssage "Expected A, B or C at potision N", omitting
the fact that we've expected characters with which A, B and C begin.