Make define-cfn accept :extern for forward decl generation
There is no way currently to write a forward function declaration in
sexp. One would have to fall back to write a C fragment, which is not
as nice.

define-cfn seems like the best place to add support for this since it
already knows how to generate forward declaration and we do not need
to introduce another form just for this.

A few extensions are updated to take advantage of this change.