Add new CiSE form to generate struct/union
This form follows closely define-cfn where the "type" is now either
union or struct (default to struct) and the only supported qualifier
is :typedef. This should be enough to define most data structures in
CiSE.

Nested struct definition is not supported (not worth it). Anonymous
struct is not supported. "typedef struct foo bar;" is also not
supported.