Avoid argument shifting in TAIL-APPLY insn
Instead of shift the arguments already in stack, we can just
push the restarg and move ARGP pointer.

The effect of this optimization is subtle, supposedly the
stack content is already on cache, and other operations related
to the call is more significant.  In a micro benchmark tailored
to emphasize this effect, we see 1-2% of improvements.
It doesn't harm though, and it simplifies the code, so why not?