Avoid printing object in the InstanceSlotRef slot unbound message
If write-object method is defined on the object and the method dereferences
the slot which is unbound, then we enter InstanceSlotRef again and
causes double fault.

We now take the same approach as the ordinary slot-ref: just show the
class of the offending object.

Related: https://github.com/shirok/Gauche/issues/488