Partial SRE support with regexp-{parse,compile}-sre
This adds a regexp-parse-sre to convert SRFI-115 SRE to Regexp AST
(and regexp-compile-sre finishes the rest). Support is not complete
because the AST may need some extension to cover full SRE syntax. See
FIXME comments for more info.

This is experimental (and hopefully not that buggy). Once we cover full
SRE syntax, we could start thinking about real SRFI-115 support.

AST tree is extended to handle begin and end of word separately.
regexp-unparse is updated so that it does not throw error on bow/eow
because we may want print/trace the regexp objects (especially in
error context). The result regexp string will of course not parse.