sre: add regexp-unparse-sre
This is basically SRFI-115's regexp->sre.

Note that we will produce native <char-set> in the SRE. So if you
print (or save) it, you'll get Gauche's #[...] read syntax, which is not
really portable.

The good news is, SRFI-115 says nothing about portability (while it does
mention that for char-set->sre). So this is probably OK. The user could
do a recursive map of char-set->sre to get portable SRE.