regexp: add multi-line mode
The main reason for this is because SRE supports multi-line matching (it
actually just does not say anything about an SRE having to match just
one line).

In multi-line mode, bol and eol match both begin/end of the string _and_
each line. The BOL_ANCHORED optimization needs to turn off to enable
back tracking.

regexp-parse, regexp-compile, regexp-compile-sre and string-regexp are
updated to take a new :multi-line keyword for this. By default multi-line
is off, except for regexp-compile-sre.