regexp: extend 'rxmatch' to search in a substring
This is needed by SRFI-115 which provides procedure like

    (regexp-search re str [start [end]]) => regexp-match-or-false

Simulating it by calling (substring str start end) first and passing the
substring to rxmatch is not enough because the start/end offsets in the
match objects will be relative to the substring instead of the full
string (chibi goes with the latter).

The new start is now assigned to ctx->input to be the real beginning of
the search string. make_match() still works on the original string
object, so all offsets are relative to full string.