Rename SCM_CHAR_SET_DIGIT to SCM_CHAR_SET_ASCII_DIGIT
This is the first step to change char-set:digit to cover full unicode
range. Many other places will need to stay with ASCII range though. To
make sure we spot all of those places, rename the DIGIT charset id,
forcing us to fix (and see) all the places that use it.

As shown in the diff, SCM_CHAR_SET_DIGIT is used for

- Charset syntax #[\d] #[\D] and #[[:digit:]]
- Regular expressions \d and \D
- SRFI-14 char-set:digit and char-set:letter+digit

Only the last item will be redefined in the next patch to cover Unicode
range. The rest stays with ASCII digits.