srfi-115: fix regexp-replace taking a single symbol or number
The underlying regexp-replace code in librx.scm expects substitution to
be either a procedure or a list of things. The SRFI-115 wrapper does not
wrap symbols or numbers properly (it does wrap strings), and will
trigger an error:

    gosh[r7rs.user]$ (regexp-replace "abc" "abcdefghi" 1)
    *** ERROR: pair required, but got 1

Fix this.

Also catch the ambiguous case where the user has a capture group named
'pre or 'post and also uses 'pre or 'post in regexp-replace. We can't be
sure if the user wants a capture group as the substitution, or
regexp-{before,after}.

Raise an error (and the user may need to rename their capture
groups). The current behavior is always treat 'pre/'post as
regexp-before/after, which could be unexpected.