chibi show: avoid chibi environment monad
The environment monad in chibi is created by a macro provided by '(chibi
environment monad)' but that macro does not work in Gauche. Problem with
rtd-constructor, do manual expansion (see #532).

Avoid that macro by providing the expanded version of the monad. In the
future this code is probably replaced with srfi-165.