libmacbase.scm: keep all macro tracing messages in stderr
When *trace-macro* is set to #t (a wonderful thing to debug, should be
documented!) it prints the sexp body to stderr, the header though is to
stdout. This does not work well when I pipe both to less and
the (buffered) headers are only flushed out at the end, disconnected
from all the sexp.