Workaround the issue with -rpath and $LD_LIBRARY_PATH
When gosh is first built, it it built with -rpath pointing to
$(top_srcdir)/src so that the latest libgauche.so is always used,
ignoring the same .so in system.

This should work fine when -rpath creates RPATH section in the ELF. But
since gcc 7.3, the linker now creates RUNPATH section instead. The
difference between these is search order in ld.so: the linker searches
RPATH first, then LD_LIBRARY_PATH and then RUN_PATH.

So if LD_LIBRARY_PATH is already set and a new compiler is used on
Linux, the -rpath setting above is ignored. As a result, ld.so could
load the (outdated) libgauche.so from LD_LIBRARY_PATH instead. This
could lead to strange crashes.

To work around this issue, detect LD_LIBRARY_PATH, adjust it and exec()
again. The WA is limited to GNU/Linux environment because it probably
only happens in this environment.

The workaround in theory should be applied regardless of -ftest. But
that's wrong because at installation time, we relink and remove -rpath,
which means no workaround needed anymore. We can't detect that in C. So
keep it to -ftest only, even if it means running gosh from source tree
without -ftest in this case could segfault.

Related to #556